requirements.txt
arrow
Django<5.0
django-celery-beat
django-celery-results
django-cors-headers
django-filter
django-mptt
django-mptt-admin
django-postgrespool2
django-redis==5.4.0
django-rest-captcha==0.3.0
django-timezone-field
djangorestframework==3.14.0
drf-yasg
cryptography
gunicorn
importlib-metadata==4.13.0
importlib-resources==5.7.1
ipip-ipdb==1.6.1
ipython
Jinja2
Pillow==9.5.0
psycopg2-binary
PyJWT<2.0
PyMySQL
pypdf==3.17.4
pypinyin
pyquery<1.5
python-docx
pytz
PyYAML
redis
reportlab
requests
Shapely<1.9
wechatpy<1.9
